修改database和script的entity结构，给service接口添加事务注解
