+++ src/main/java/com/creams/temo/entity/database/Script.java
@@ -1,4 +1,4 @@
-package com.creams.temo.entity.project.response;
+package com.creams.temo.entity.database;
 
 import com.baomidou.mybatisplus.annotation.IdType;
 import com.baomidou.mybatisplus.annotation.TableField;
@@ -7,13 +7,9 @@ import com.baomidou.mybatisplus.annotation.TableName;
 import io.swagger.annotations.ApiModelProperty;
 import lombok.Data;
 
-/**
- * @author jelly
- * @since 1.0
- */
 @Data
 @TableName(value = "script")
-public class ScriptResponse {
+public class Script {
 
     @ApiModelProperty("主键")
     @TableId(value = "id", type = IdType.AUTO)
@@ -42,6 +38,7 @@ public class ScriptResponse {
     @ApiModelProperty("修改时间")
     @TableField(value = "update_time")
     private String updateTime;
+
 
 
 }
