+++ src/main/java/com/creams/temo/entity/database/request/DatabaseRequest.java
@@ -1,23 +1,15 @@
-package com.creams.temo.entity.project;
+package com.creams.temo.entity.database.request;
 
-import com.baomidou.mybatisplus.annotation.IdType;
 import com.baomidou.mybatisplus.annotation.TableField;
-import com.baomidou.mybatisplus.annotation.TableId;
-import com.baomidou.mybatisplus.annotation.TableName;
 import io.swagger.annotations.ApiModelProperty;
 import lombok.Data;
-import java.sql.Timestamp;
 
 @Data
-@TableName(value = "db")
-public class Database {
+public class DatabaseRequest {
 
-    @ApiModelProperty("主键")
-    @TableId(value = "id", type = IdType.AUTO)
-    private Integer id;
-
-    @ApiModelProperty("数据库id")
+    @ApiModelProperty(value = "数据库id",hidden = true)
     @TableField(value = "db_id")
+
     private String dbId;
 
     @ApiModelProperty("数据库名称")
@@ -28,7 +20,7 @@ public class Database {
     @TableField(value = "host")
     private String host;
 
-    @ApiModelProperty("断口号")
+    @ApiModelProperty("端口号")
     @TableField(value = "port")
     private Integer port;
 
@@ -43,7 +35,5 @@ public class Database {
     @ApiModelProperty("数据库名称")
     @TableField(value = "db_library_name")
     private String dbLibraryName;
-
-
 
 }
