+++ src/main/java/com/creams/temo/entity/database/request/ScriptRequest.java
@@ -1,10 +1,7 @@
-package com.creams.temo.entity.project.request;
+package com.creams.temo.entity.database.request;
 
-import com.baomidou.mybatisplus.annotation.IdType;
 import com.baomidou.mybatisplus.annotation.TableField;
-import com.baomidou.mybatisplus.annotation.TableId;
 import io.swagger.annotations.ApiModelProperty;
-import io.swagger.annotations.ApiParam;
 import lombok.Data;
 
 import java.sql.Timestamp;
